# api/docs/swagger.yaml
basePath: /
definitions:
...
  budgeting_service.Categories:
    properties:
      categories:
        items:
          $ref: '#/definitions/budgeting_service.Category'
        type: array
    type: object
  budgeting_service.Category:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  budgeting_service.CreateCategory:
    properties:
      name:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  budgeting_service.CreateGoal:
    properties:
      current_amount:
        type: number
      deadline:
        type: string
      name:
        type: string
      status:
        type: string
      target_amount:
        type: number
      user_id:
        type: string
    type: object
...
  budgeting_service.Goal:
    properties:
      created_at:
        type: string
      current_amount:
        type: number
      deadline:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: string
      target_amount:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  budgeting_service.Goals:
    properties:
      goals:
        items:
          $ref: '#/definitions/budgeting_service.Goal'
        type: array
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
...
info:
  contact: {}
  description: Something big
  title: MoneyMate
  version: "1.0"
paths:
...
  /categories/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category retrieved successfully
          schema:
            $ref: '#/definitions/budgeting_service.Category'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - categories
  /categories/{id}/delete:
    delete:
      consumes:
      - application/json
      description: Deletes a category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted successfully
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - categories
  /categories/{id}/update:
    put:
      consumes:
      - application/json
      description: Updates an existing category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category Update Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/budgeting_service.Category'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated successfully
          schema:
            $ref: '#/definitions/budgeting_service.Category'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - categories
  /categories/all:
    get:
      consumes:
      - application/json
      description: Retrieves all categories
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Limit number
        in: query
        name: limit
        required: true
        type: integer
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Name
        in: query
        name: name
        type: string
      - description: Type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Categories retrieved successfully
          schema:
            $ref: '#/definitions/budgeting_service.Categories'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - categories
  /categories/create:
    post:
      consumes:
      - application/json
      description: Creates a new category
      parameters:
      - description: Category Creation Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/budgeting_service.CreateCategory'
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
          schema:
            $ref: '#/definitions/budgeting_service.Category'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - categories
  /goals/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a goal by ID
      parameters:
      - description: Goal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Goal retrieved successfully
          schema:
            $ref: '#/definitions/budgeting_service.Goal'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - goals
  /goals/{id}/delete:
    delete:
      consumes:
      - application/json
      description: Deletes a goal by ID
      parameters:
      - description: Goal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Goal deleted successfully
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - goals
  /goals/{id}/update:
    put:
      consumes:
      - application/json
      description: Updates an existing goal
      parameters:
      - description: Goal ID
        in: path
        name: id
        required: true
        type: string
      - description: Goal Update Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/budgeting_service.Goal'
      produces:
      - application/json
      responses:
        "200":
          description: Goal updated successfully
          schema:
            $ref: '#/definitions/budgeting_service.Goal'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - goals
  /goals/all:
    get:
      consumes:
      - application/json
      description: Retrieves all goals
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Limit number
        in: query
        name: limit
        required: true
        type: integer
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Name
        in: query
        name: name
        type: string
      - description: Target Amount
        in: query
        name: target_amount
        type: number
      - description: Current Amount
        in: query
        name: current_amount
        type: number
      - description: Deadline
        in: query
        name: deadline
        type: string
      - description: Status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Goals retrieved successfully
          schema:
            $ref: '#/definitions/budgeting_service.Goals'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - goals
  /goals/create:
    post:
      consumes:
      - application/json
      description: Creates a new goal
      parameters:
      - description: Goal Creation Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/budgeting_service.CreateGoal'
      produces:
      - application/json
      responses:
        "201":
          description: Goal created successfully
          schema:
            $ref: '#/definitions/budgeting_service.Goal'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      tags:
      - goals
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"